type to article visibility

+++ app/Http/Controllers/TypeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Type;
+use App\Models\Article;
 use App\Http\Requests\StoreTypeRequest;
 use App\Http\Requests\UpdateTypeRequest;
 use Illuminate\Http\Request;
@@ -17,7 +18,8 @@ class TypeController extends Controller
     public function index()
     {
         $type = Type::all();
-        return view("type.index", ['types' => $type]);
+        $article = Article::all();
+        return view("type.index", ['types' => $type, 'articles' => $article]);
     }
 
     /**
@@ -27,7 +29,7 @@ class TypeController extends Controller
      */
     public function create()
     {
-        //
+        return view("create.index");
     }
 
     /**
@@ -36,9 +38,14 @@ class TypeController extends Controller
      * @param  \App\Http\Requests\StoreTypeRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StoreTypeRequest $request)
+    public function store(Request $request)
     {
-        //
+        $type = new Type;
+        $type->title = $request->type_title;
+        $type->description = $request->type_description;
+
+        $type->save();
+        return redirect()->route('type.index');
     }
 
     public function storeAjax(Request $request)
@@ -46,14 +53,15 @@ class TypeController extends Controller
         $type = new Type;
         $type->title = $request->type_title;
         $type->description = $request->type_description;
-        
+
         $type->save();
 
         $type_array = array(
             'successMessage' => "type stored succesfuly",
             'typeId' => $type->id,
             'typeTitle' => $type->title,
-            'typeDescription' => $type->description,            
+            'typeDescription' => $type->description,
+            'typeHasArticles' => count($type->typeArticles),
         );
 
         $json_response = response()->json($type_array);
@@ -72,7 +80,8 @@ class TypeController extends Controller
         //
     }
 
-    public function showAjax(Type $type) {
+    public function showAjax(Type $type)
+    {
 
         $type_array = array(
             'successMessage' => "Type retrieved succesfuly",
@@ -81,7 +90,7 @@ class TypeController extends Controller
             'typeDescription' => $type->description,
         );
 
-        $json_response =response()->json($type_array); 
+        $json_response = response()->json($type_array);
 
         return $json_response;
     }
@@ -113,7 +122,7 @@ class TypeController extends Controller
     {
         $type->title = $request->type_title;
         $type->description = $request->type_description;
-    
+
         $type->save();
 
         $type_array = array(
@@ -123,7 +132,7 @@ class TypeController extends Controller
             'typeDescription' => $type->description,
         );
 
-        $json_response =response()->json($type_array); 
+        $json_response = response()->json($type_array);
 
         return $json_response;
     }
@@ -142,15 +151,20 @@ class TypeController extends Controller
 
     public function destroyAjax(Type $type)
     {
-        $type->delete();
 
-        $success_array = array(
-            'successMessage' => $type->title. " type deleted successfuly"
-        );
 
-        $json_response =response()->json($success_array);
+        if (count($type->typeArticles) > 0) {
+            $response_array = array(
+                'errorMessage' => $type->title . " type can't be deleted, bacouse it has ".count($type->typeArticles)." articles");
+        } else {
+
+            $type->delete();
+            $response_array = array(
+                'successMessage' => $type->title . " type deleted successfuly"
+            );
+        }
+        $json_response = response()->json($response_array);
 
         return $json_response;
-
     }
 }
